[txn_generator] Separate out transaction_generator_lib
So we can reuse it executor benchmark
