+++ crates/transaction-emitter-lib/src/lib.rs
@@ -7,7 +7,6 @@ mod args;
 mod cluster;
 pub mod emitter;
 mod instance;
-mod transaction_generator;
 mod wrappers;
 
 // These are the top level things you should need to run the emitter.
@@ -17,7 +16,6 @@ pub use cluster::Cluster;
 pub use emitter::{
     query_sequence_number, query_sequence_numbers,
     stats::{TxnStats, TxnStatsRate},
-    EmitJob, EmitJobMode, EmitJobRequest, EmitModeParams, TransactionType, TxnEmitter,
+    EmitJob, EmitJobMode, EmitJobRequest, EmitModeParams, TxnEmitter,
 };
-pub use transaction_generator::EntryPoints;
 pub use wrappers::{emit_transactions, emit_transactions_with_cluster};
