+++ crates/transaction-generator-lib/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "aptos-transaction-emitter-lib"
-description = "Library for Aptos transaction emitter"
+name = "aptos-transaction-generator-lib"
+description = "Library for Aptos transaction generator"
 version = "0.0.0"
 
 # Workspace inherited keys
@@ -24,7 +24,6 @@ aptos-infallible = { workspace = true }
 aptos-logger = { workspace = true }
 aptos-rest-client = { workspace = true }
 aptos-sdk = { workspace = true }
-aptos-transaction-generator-lib = { workspace = true }
 async-trait = { workspace = true }
 clap = { workspace = true }
 futures = { workspace = true }
