+++ crates/transaction-generator-lib/src/accounts_pool_wrapper.rs
@@ -1,7 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::transaction_generator::{TransactionGenerator, TransactionGeneratorCreator};
+use crate::{TransactionGenerator, TransactionGeneratorCreator};
 use aptos_infallible::RwLock;
 use aptos_logger::{sample, sample::SampleRate, warn};
 use aptos_sdk::types::{transaction::SignedTransaction, LocalAccount};
