+++ crates/transaction-generator-lib/src/p2p_transaction_generator.rs
@@ -1,6 +1,6 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
-use crate::transaction_generator::{TransactionGenerator, TransactionGeneratorCreator};
+use crate::{TransactionGenerator, TransactionGeneratorCreator};
 use aptos_infallible::RwLock;
 use aptos_sdk::{
     move_types::account_address::AccountAddress,
