+++ testsuite/forge/src/lib.rs
@@ -14,6 +14,7 @@ pub use runner::*;
 
 mod backend;
 pub use aptos_transaction_emitter_lib::*;
+pub use aptos_transaction_generator_lib::*;
 pub use backend::*;
 
 mod report;
